+++ Code/Views/ATLMessageInputToolbar.h
@@ -90,7 +90,7 @@ extern NSString *const ATLMessageInputToolbarAccessibilityLabel;
 
 /**
  @abstract The left accessory button for the view. 
- @discussion By default, the button displays a camera icon.
+ @discussion By default, the button displays a camera icon. If set to `nil` the `textInputView` will expand to the left edge of the toolbar.
  */
 @property (nonatomic) UIButton *leftAccessoryButton;
  
@@ -99,6 +99,25 @@ extern NSString *const ATLMessageInputToolbarAccessibilityLabel;
  @discussion By default, the button displays the text "SEND".
  */
 @property (nonatomic) UIButton *rightAccessoryButton;
+
+/**
+ @abstract The image displayed on left accessory button.
+ @default A `camera` icon.
+ */
+@property (nonatomic) UIImage *leftAccessoryImage;
+
+/**
+ @abstract The image displayed on right accessory button.
+ @default A `location` icon.
+ */
+@property (nonatomic) UIImage *rightAccessoryImage;
+
+/**
+ @abstract Determines whether or not the right accessory button displays an icon. 
+ @disucssion If NO, the right accessory button will display the text `SEND` at all times.
+ @default YES
+ */
+@property(nonatomic) BOOL displaysRightAccessoryImage;
 
 /**
  @abstract An automatically resizing message composition field.